refactor: Remove unused components and files to streamline the project structure
